完成觸發邏輯與unit test

+++ internal/structs.go
@@ -4,7 +4,7 @@ import "time"
 
 type AlarmCacher struct {
 	Object                    string    `json:"object"`
-	EventID                   int64     `json:"event_id"`
+	EventID                   int       `json:"event_id"`
 	AlarmCategoryCurrent      string    `json:"alar_category"`
 	AlarmCategoryOrderCurrent int       `json:"alar_category_order_current"`
 	AlarmCategoryHigh         string    `json:"alar_category_high"`
